Renamed playlist API for consistency
